Add unit tests for the connection limit
